+++ packages/student-witness/src/app.js
@@ -1,3 +1,0 @@
-import { lib } from './utils/lib';
-
-console.log('heheda', lib());